pyproject.toml
[tool.poetry]
name = "heartlink"
version = "0.1.0"
description = ""
authors = ["Ilya <ilya925512@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.112.2"}
asyncpg = "^0.29.0"
python-dotenv = "^1.0.1"
asyncio = "^3.4.3"
sqlalchemy = "^2.0.32"
alembic = "^1.13.2"
pyjwt = "^2.9.0"
bcrypt = "^4.2.0"
aiobotocore = "^2.14.0"
greenlet = "^3.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
